Fix tests by passing in an empty CliContext
