+++ go/cmd/dolt/commands/sql_test.go
@@ -17,8 +17,6 @@ package commands
 import (
 	"context"
 	"fmt"
-	"github.com/dolthub/dolt/go/cmd/dolt/cli"
-	"github.com/dolthub/dolt/go/libraries/utils/argparser"
 	"testing"
 
 	"github.com/dolthub/go-mysql-server/sql"
@@ -45,7 +43,7 @@ import (
 
 var tableName = "people"
 
-var stubCliCtx = buildMockCliContext()
+var stubCliCtx = BuildEmptyCliContext()
 
 // Smoke test: Console opens and exits
 func TestSqlConsole(t *testing.T) {
@@ -57,7 +55,7 @@ func TestSqlConsole(t *testing.T) {
 		args := []string{}
 		commandStr := "dolt sql"
 
-		result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+		result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 		assert.Equal(t, 0, result)
 	})
 
@@ -85,7 +83,7 @@ func TestSqlBatchMode(t *testing.T) {
 			args := []string{"-b", "-q", test.query}
 
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 		})
 	}
@@ -124,7 +122,7 @@ func TestSqlSelect(t *testing.T) {
 			args := []string{"-q", test.query}
 
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 		})
 	}
@@ -150,7 +148,7 @@ func TestSqlShow(t *testing.T) {
 			args := []string{"-q", test.query}
 
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 		})
 	}
@@ -183,7 +181,7 @@ func TestCreateTable(t *testing.T) {
 
 			args := []string{"-q", test.query}
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 
 			working, err = dEnv.WorkingRoot(context.Background())
@@ -223,7 +221,7 @@ func TestShowTables(t *testing.T) {
 
 			args := []string{"-q", test.query}
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 		})
 	}
@@ -254,7 +252,7 @@ func TestAlterTable(t *testing.T) {
 
 			args := []string{"-q", test.query}
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 		})
 	}
@@ -281,7 +279,7 @@ func TestDropTable(t *testing.T) {
 
 			args := []string{"-q", test.query}
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(context.TODO(), commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 		})
 	}
@@ -400,7 +398,7 @@ func TestInsert(t *testing.T) {
 			args := []string{"-q", test.query}
 
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(ctx, commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(ctx, commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 
 			if result == 0 {
@@ -481,7 +479,7 @@ func TestUpdate(t *testing.T) {
 			args := []string{"-q", test.query}
 
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(ctx, commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(ctx, commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 
 			if result == 0 {
@@ -556,7 +554,7 @@ func TestDelete(t *testing.T) {
 
 			ctx := context.Background()
 			commandStr := "dolt sql"
-			result := SqlCmd{}.Exec(ctx, commandStr, args, dEnv, &stubCliCtx)
+			result := SqlCmd{}.Exec(ctx, commandStr, args, dEnv, stubCliCtx)
 			assert.Equal(t, test.expectedRes, result)
 
 			if result == 0 {
@@ -594,18 +592,4 @@ func TestCommitHooksNoErrors(t *testing.T) {
 			t.Errorf("expected LogHook, found: %s", h)
 		}
 	}
-}
-
-type cliCtx struct{}
-
-func (c cliCtx) GlobalArgs() *argparser.ArgParseResults {
-	ap := argparser.NewArgParserWithMaxArgs("sql test", 0)
-	apr, _ := ap.Parse(make([]string, 0))
-	return apr
-}
-
-var _ cli.CliContext = cliCtx{}
-
-func buildMockCliContext() cli.CliContext {
-	return cliCtx{}
 }
