+++ go/libraries/doltcore/doltdb/foreign_key_test.go
@@ -37,6 +37,8 @@ func TestForeignKeys(t *testing.T) {
 	}
 }
 
+var fkCliCtx = commands.BuildEmptyCliContext()
+
 func TestForeignKeyErrors(t *testing.T) {
 	skipNewFormat(t)
 	cmds := []testCommand{
@@ -49,13 +51,13 @@ func TestForeignKeyErrors(t *testing.T) {
 	dEnv := dtestutils.CreateTestEnv()
 
 	for _, c := range cmds {
-		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, nil)
+		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, fkCliCtx)
 		require.Equal(t, 0, exitCode)
 	}
 
-	exitCode := commands.SqlCmd{}.Exec(ctx, commands.SqlCmd{}.Name(), []string{"-q", `ALTER TABLE test MODIFY v1 INT;`}, dEnv, nil)
+	exitCode := commands.SqlCmd{}.Exec(ctx, commands.SqlCmd{}.Name(), []string{"-q", `ALTER TABLE test MODIFY v1 INT;`}, dEnv, fkCliCtx)
 	require.Equal(t, 1, exitCode)
-	exitCode = commands.SqlCmd{}.Exec(ctx, commands.SqlCmd{}.Name(), []string{"-q", `ALTER TABLE test2 MODIFY v1 INT;`}, dEnv, nil)
+	exitCode = commands.SqlCmd{}.Exec(ctx, commands.SqlCmd{}.Name(), []string{"-q", `ALTER TABLE test2 MODIFY v1 INT;`}, dEnv, fkCliCtx)
 	require.Equal(t, 1, exitCode)
 }
 
@@ -96,11 +98,11 @@ func testForeignKeys(t *testing.T, test foreignKeyTest) {
 	dEnv := dtestutils.CreateTestEnv()
 
 	for _, c := range fkSetupCommon {
-		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, nil)
+		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, gcCliCtx)
 		require.Equal(t, 0, exitCode)
 	}
 	for _, c := range test.setup {
-		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, nil)
+		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, gcCliCtx)
 		require.Equal(t, 0, exitCode)
 	}
