+++ go/libraries/doltcore/doltdb/gc_test.go
@@ -55,6 +55,8 @@ type gcTest struct {
 	postGCFunc func(ctx context.Context, t *testing.T, ddb *doltdb.DoltDB, prevRes interface{})
 }
 
+var gcCliCtx = commands.BuildEmptyCliContext()
+
 var gcTests = []gcTest{
 	{
 		name: "gc test",
@@ -113,7 +115,7 @@ func testGarbageCollection(t *testing.T, test gcTest) {
 	defer dEnv.DoltDB.Close()
 
 	for _, c := range gcSetupCommon {
-		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, nil)
+		exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, gcCliCtx)
 		require.Equal(t, 0, exitCode)
 	}
 
@@ -121,7 +123,7 @@ func testGarbageCollection(t *testing.T, test gcTest) {
 	for _, stage := range test.stages {
 		res = stage.preStageFunc(ctx, t, dEnv.DoltDB, res)
 		for _, c := range stage.commands {
-			exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, nil)
+			exitCode := c.cmd.Exec(ctx, c.cmd.Name(), c.args, dEnv, gcCliCtx)
 			require.Equal(t, 0, exitCode)
 		}
 	}
